#601 fix navbar in syncdocs

+++ Portal/utils/SyncDocs/Program.cs
@@ -13,6 +13,7 @@ using Microsoft.Extensions.Options;
 
 const string SIDEBAR = "_sidebar.md";
 const string SIDEBAR_META = "_sidebar.md.yaml";
+const string NAVBAR = "_navbar.md";
 
 var builder = new ConfigurationBuilder()
     .SetBasePath(Directory.GetCurrentDirectory())
@@ -148,7 +149,7 @@ static async Task CleanupTargetDirectory(ConfigParams configParams, string path)
     await IteratePath(Path.Combine(path, configParams.Target), s => false, f => { folders.Add(f); return Task.CompletedTask; }, f => { files.Add(f); return Task.CompletedTask; });
     foreach (var item in files.Where(f => Path.GetExtension(f) == ".md"))
     {   
-        if (MarkdownDocumentationService.CheckIfDraft(item))
+        if (MarkdownDocumentationService.CheckIfDraft(item) && !string.Equals(Path.GetFileName(item), NAVBAR, StringComparison.InvariantCultureIgnoreCase))
         {
             Console.WriteLine($"Deleting draft file {item}");
             File.Delete(item);
@@ -167,7 +168,7 @@ static async Task CleanupTargetDirectory(ConfigParams configParams, string path)
 static Func<string, bool> BuildExcluder(ConfigParams config)
 {
     var excludeSet = config.GetExcludedFolders().Select(s => s.ToLower()).ToHashSet();
-    return n => excludeSet.Contains(n.ToLower());
+    return n => excludeSet.Contains(n.ToLower()) || string.Equals(Path.GetFileName(n),NAVBAR,StringComparison.InvariantCultureIgnoreCase);
 }
 
 static int FolderDepth(string path)
